[#51] refactor : 애플리케이션 repository 메서드 시그니처 추상화
