+++ src/main/java/depromeet/lessonfour/server/activityrecord/infra/repository/ActivityRecordRepositoryImpl.java
@@ -17,7 +17,7 @@ public class ActivityRecordRepositoryImpl implements ActivityRecordRepository {
   private final QueryDslActivityRecordRepository queryDslActivityRecordRepository;
 
   @Override
-  public Optional<ActivityRecord> findByUserIdAndOccurredAt(Long userId, LocalDate date) {
+  public Optional<ActivityRecord> findByUserAndDate(Long userId, LocalDate date) {
     return queryDslActivityRecordRepository.findByUserIdAndOccurredAt(userId, date);
   }
 
@@ -27,13 +27,12 @@ public class ActivityRecordRepositoryImpl implements ActivityRecordRepository {
   }
 
   @Override
-  public boolean existsByUserIdAndActivityAt(Long userId, LocalDate activityAt) {
+  public boolean existsByUserAndDate(Long userId, LocalDate activityAt) {
     return queryDslActivityRecordRepository.existsByUserIdAndActivityAt(userId, activityAt);
   }
 
   @Override
-  public Optional<ActivityRecord> findByUser_IdAndIdAndIsDeletedFalse(
-      Long userId, Long activityRecordId) {
+  public Optional<ActivityRecord> findByUserAndId(Long userId, Long activityRecordId) {
     return jpaActivityRecordRepository.findByUser_IdAndIdAndIsDeletedFalse(
         userId, activityRecordId);
   }
